+++ ariadne_graphql_proxy/contrib/imgix/__init__.py
@@ -1,0 +1,3 @@
+from .query_params_resolver import get_query_params_resolver
+
+__all__ = ["get_query_params_resolver"]