+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `CloudflareCacheBackend`.
 - Added `DynamoDBCacheBackend`.
 - Changed `QueryFilter` and `root_resolver` to split variables between schemas.
+- Added `insert_field` utility to `ProxySchema`. Added `get_query_params_resolver` as factory for `imgix` resolvers. 
 
 
 ## 0.1.0 (2023-06-13)